+++ src/include/utils.h
@@ -10,41 +10,48 @@
     #define WIN32_LEAN_AND_MEAN /* compact Win32 to only common utilities */
 #endif
 
-#ifndef BENJI_NO_SERVER_INCLUDES
-    #if defined(_WIN32)
-        #define BENJI_ABI __cdecl
-
-        #pragma comment(lib, "ws2_32.lib")
-
-        #include <winsock2.h>
-        #include <ws2tcpip.h>
-
-        typedef SOCKET BENJI_SOCKET;
-
-        BENJI_ABI void winsock_init();
-        BENJI_ABI void winsock_cleanup();
-    #elif defined(__linux__)
-        #define BENJI_ABI
-
-        #include <sys/socket.h>
-        #include <arpa/inet.h>
-        #include <unistd.h>
-
-        typedef unsigned long long BENJI_SOCKET;
-    #endif
-
+#ifdef NO_ERROR
+    #define BENJI_NO_ERROR NO_ERROR
+#else
     #define BENJI_NO_ERROR 0
-    #define BENJI_SOCKET_ERROR -1
-
-    #define BENJI_MAX_SOCK_CONNS 5 // maximum number of tries to connect to a socket
-
-    BENJI_ABI BENJI_SOCKET create_socket();
-    BENJI_ABI void close_socket(BENJI_SOCKET sock);
-
-    BENJI_ABI void stop(const size_t exit_code);
 #endif
 
-#ifndef BENJI_NO_SYS_INFO_INCLUDES
+// #ifdef BENJI_USE_SERVER_UTILS
+//     #if defined(_WIN32)
+//         #define BENJI_SC_ABI __cdecl /* Benji Server Call */
+
+//         #pragma comment(lib, "ws2_32.lib")
+
+//         #include <winsock2.h>
+//         #include <ws2tcpip.h>
+
+//         typedef SOCKET BENJI_SOCKET;
+
+//         #define BENJI_SOCKET_ERROR SOCKET_ERROR
+
+//         BENJI_SC_ABI void winsock_init();
+//         BENJI_SC_ABI void winsock_cleanup();
+//     #elif defined(__linux__)
+//         #define BENJI_SC_ABI
+
+//         #include <sys/socket.h>
+//         #include <arpa/inet.h>
+//         #include <unistd.h>
+
+//         typedef unsigned long long BENJI_SOCKET;
+
+//         #define BENJI_SOCKET_ERROR -1
+//     #endif
+
+//     #define BENJI_MAX_SOCK_CONNS 5 /* maximum number of tries to connect to a socket */
+
+//     BENJI_SC_ABI BENJI_SOCKET create_socket();
+//     BENJI_SC_ABI void close_socket(BENJI_SOCKET sock);
+
+//     BENJI_SC_ABI void terminate(const size_t exit_code);
+// #endif
+
+#ifdef BENJI_USE_SYS_INFO_UTILS
     #if defined(_WIN32)
         enum BENJI_WMI_TYPE {
             WMI_WIN32_PROCESSOR,
