+++ src/server.c
@@ -1,6 +1,6 @@
 #include "include/server.h"
 
-BENJI_ABI BENJI_SOCKET create_server() {
+BENJI_SC_ABI BENJI_SOCKET create_server() {
     struct sockaddr_in server_address;
 
     server_address.sin_family = AF_INET; // ipv4 address family
@@ -18,13 +18,13 @@ BENJI_ABI BENJI_SOCKET create_server() {
     if (bind(server_socket, (struct sockaddr*) &server_address, sizeof(server_address)) == BENJI_SOCKET_ERROR) {
         printf("Failed to bind to server address\n");
 
-        stop(EXIT_FAILURE);
+        terminate(EXIT_FAILURE);
     }
 
     if (listen(server_socket, BENJI_MAX_SOCK_CONNS) == BENJI_SOCKET_ERROR) {
         printf("Failed to put into listening mode\n");
 
-        stop(EXIT_FAILURE);
+        terminate(EXIT_FAILURE);
     }
 
     printf("Success\n\n");
